Build with GCC

+++ koro_migrator/create.c
@@ -3,12 +3,16 @@
 static char *koro_getDatabaseName() {
   char *env = koro_getFlavor();
   KoroConfig *config = koro_readConfig();
-  if (config == NULL) exit(KORO_CONFIG_MISSING);
+  if (config == NULL) {
+    NO_DB_CONFIG_MSG
+    return NULL;
+  }
   KoroEnvironmentConfig *envConfig = koro_findEnvConfig(config, env);
   free(env);
   if (envConfig == NULL) {
     koro_freeConfig(config);
-    exit(KORO_CONFIG_MISSING);
+    NO_DB_CONFIG_FOR_ENV_MSG
+    return NULL;
   }
 
   char *name = calloc(sizeof(char), strlen(envConfig->name) + 1);
@@ -27,6 +31,8 @@ char koro_isDatabaseDrop(char *str) {
 
 int koro_createDatabase() {
   char *name = koro_getDatabaseName();
+  if (name == NULL)
+    return KORO_CONFIG_MISSING;
   char *buffer = calloc(sizeof(char), strlen("CREATE DATABASE ") + strlen(name) + 1);
   strcat(buffer, "CREATE DATABASE ");
   strcat(buffer, name);
@@ -41,6 +47,9 @@ int koro_createDatabase() {
 
 int koro_dropDatabase() {
   char *name = koro_getDatabaseName();
+  if (name == NULL)
+    return KORO_CONFIG_MISSING;
+
   char *buffer = calloc(sizeof(char), strlen("DROP DATABASE ") + strlen(name) + 1);
   strcat(buffer, "DROP DATABASE ");
   strcat(buffer, name);
